#168 [feat] 네트워크 에러 뷰에서의 로직 수정
