+++ app/src/main/java/hous/release/android/di/RetrofitModule.kt
@@ -3,8 +3,6 @@ package hous.release.android.di
 import android.content.Context
 import android.content.Intent
 import android.content.SharedPreferences
-import android.net.ConnectivityManager
-import android.net.NetworkCapabilities
 import android.os.Handler
 import android.os.Looper
 import com.google.gson.GsonBuilder
@@ -18,6 +16,7 @@ import hous.release.android.R
 import hous.release.android.presentation.login.LoginActivity
 import hous.release.android.presentation.network_error.NetworkErrorActivity
 import hous.release.android.util.ToastMessageUtil
+import hous.release.android.util.extension.isNetworkConnected
 import hous.release.data.datasource.LocalPrefTokenDataSource
 import hous.release.data.repository.RefreshRepositoryImpl.Companion.EXPIRED_REFRESH_TOKEN
 import hous.release.data.repository.RefreshRepositoryImpl.Companion.EXPIRED_TOKEN
@@ -25,7 +24,6 @@ import hous.release.domain.repository.RefreshRepository
 import kotlinx.coroutines.runBlocking
 import okhttp3.Interceptor
 import okhttp3.OkHttpClient
-import okhttp3.Response
 import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.HttpException
 import retrofit2.Retrofit
@@ -57,86 +55,73 @@ object RetrofitModule {
         localPref: SharedPreferences,
         refreshRepository: RefreshRepository,
         localPrefTokenDataSource: LocalPrefTokenDataSource
-    ): Interceptor = object : Interceptor {
-        fun isNetworkConnected(): Boolean {
-            var isConnected = false
-            val cm = context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
-            val capabilities = cm.getNetworkCapabilities(cm.activeNetwork)
-            if (capabilities != null) {
-                isConnected =
-                    capabilities.hasTransport(NetworkCapabilities.TRANSPORT_WIFI) || capabilities.hasTransport(NetworkCapabilities.TRANSPORT_CELLULAR)
-            }
-            return isConnected
-        }
-
-        override fun intercept(chain: Interceptor.Chain): Response {
-            if (!isNetworkConnected()) {
-                context.startActivity(
-                    Intent(
-                        context, NetworkErrorActivity::class.java
-                    ).apply {
-                        addFlags(Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK)
-                    }
-                )
-            }
-            val request = chain.request()
-            var response = chain.proceed(
-                request
-                    .newBuilder()
-                    .addHeader(
-                        HEADER_AUTHORIZATION,
-                        BEARER + localPrefTokenDataSource.accessToken
-                    )
-                    .addHeader(HEADER_OS_TYPE, OS_TYPE)
-                    .addHeader(HEADER_VERSION, BuildConfig.VERSION_NAME)
-                    .build()
+    ): Interceptor = Interceptor { chain ->
+        if (!context.isNetworkConnected()) {
+            context.startActivity(
+                Intent(
+                    context, NetworkErrorActivity::class.java
+                ).apply {
+                    addFlags(Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK)
+                }
             )
-            when (response.code) {
-                NETWORK_ERROR -> context.startActivity(Intent(context, NetworkErrorActivity::class.java))
-                EXPIRED_TOKEN -> {
-                    runBlocking {
-                        refreshRepository.refreshHousToken()
-                            .onSuccess {
-                                response = chain.proceed(
-                                    request
-                                        .newBuilder()
-                                        .addHeader(
-                                            HEADER_AUTHORIZATION,
-                                            BEARER + localPrefTokenDataSource.accessToken
-                                        )
-                                        .addHeader(HEADER_OS_TYPE, OS_TYPE)
-                                        .addHeader(HEADER_VERSION, BuildConfig.VERSION_NAME)
-                                        .build()
-                                )
-                            }
-                            .onFailure { throwable ->
-                                Timber.e("토큰 갱신 실패 ${throwable.message}")
-                                if (throwable is HttpException) {
-                                    when (throwable.code()) {
-                                        EXPIRED_REFRESH_TOKEN -> {
-                                            with(localPref.edit()) {
-                                                clear()
-                                                commit()
-                                            }
-                                            Handler(Looper.getMainLooper()).post(
-                                                Runnable {
-                                                    ToastMessageUtil.showToast(context, context.getString(R.string.refresh_error))
-                                                    context.startActivity(
-                                                        Intent(context, LoginActivity::class.java).apply {
-                                                            addFlags(Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK)
-                                                        }
-                                                    )
-                                                }
-                                            )
+        }
+        val request = chain.request()
+        var response = chain.proceed(
+            request
+                .newBuilder()
+                .addHeader(
+                    HEADER_AUTHORIZATION,
+                    BEARER + localPrefTokenDataSource.accessToken
+                )
+                .addHeader(HEADER_OS_TYPE, OS_TYPE)
+                .addHeader(HEADER_VERSION, BuildConfig.VERSION_NAME)
+                .build()
+        )
+        when (response.code) {
+            NETWORK_ERROR -> context.startActivity(Intent(context, NetworkErrorActivity::class.java))
+            EXPIRED_TOKEN -> {
+                runBlocking {
+                    refreshRepository.refreshHousToken()
+                        .onSuccess {
+                            response = chain.proceed(
+                                request
+                                    .newBuilder()
+                                    .addHeader(
+                                        HEADER_AUTHORIZATION,
+                                        BEARER + localPrefTokenDataSource.accessToken
+                                    )
+                                    .addHeader(HEADER_OS_TYPE, OS_TYPE)
+                                    .addHeader(HEADER_VERSION, BuildConfig.VERSION_NAME)
+                                    .build()
+                            )
+                        }
+                        .onFailure { throwable ->
+                            Timber.e("토큰 갱신 실패 ${throwable.message}")
+                            if (throwable is HttpException) {
+                                when (throwable.code()) {
+                                    EXPIRED_REFRESH_TOKEN -> {
+                                        with(localPref.edit()) {
+                                            clear()
+                                            commit()
                                         }
+                                        Handler(Looper.getMainLooper()).post(
+                                            Runnable {
+                                                ToastMessageUtil.showToast(context, context.getString(R.string.refresh_error))
+                                                context.startActivity(
+                                                    Intent(context, LoginActivity::class.java).apply {
+                                                        addFlags(Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK)
+                                                    }
+                                                )
+                                            }
+                                        )
                                     }
                                 }
                             }
-                    }
+                        }
                 }
             }
-            return response
         }
+        response
     }
 
     @Provides
