+++ app/src/main/java/hous/release/android/presentation/network_error/NetworkErrorActivity.kt
@@ -11,7 +11,9 @@ import hous.release.android.presentation.enter_room.EnterRoomActivity
 import hous.release.android.presentation.login.LoginActivity
 import hous.release.android.presentation.main.MainActivity
 import hous.release.android.presentation.tutorial.TutorialActivity
+import hous.release.android.util.ToastMessageUtil
 import hous.release.android.util.binding.BindingActivity
+import hous.release.android.util.extension.isNetworkConnected
 import hous.release.domain.entity.SplashState
 import hous.release.domain.usecase.GetSplashStateUseCase
 import javax.inject.Inject
@@ -27,16 +29,20 @@ class NetworkErrorActivity :
         checkNetworkConnect()
     }
 
-    fun checkNetworkConnect() {
+    private fun checkNetworkConnect() {
         binding.btnNetworkErrorRetry.setOnClickListener {
-            when (getSplashStateUseCase()) {
-                SplashState.TUTORIAL -> startActivity<TutorialActivity>()
-                SplashState.LOGIN -> startActivity<LoginActivity>()
-                SplashState.ENTER_ROOM -> startActivity<EnterRoomActivity>()
-                SplashState.MAIN -> startActivity<MainActivity>()
+            if (isNetworkConnected()) {
+                when (getSplashStateUseCase()) {
+                    SplashState.TUTORIAL -> startActivity<TutorialActivity>()
+                    SplashState.LOGIN -> startActivity<LoginActivity>()
+                    SplashState.ENTER_ROOM -> startActivity<EnterRoomActivity>()
+                    SplashState.MAIN -> startActivity<MainActivity>()
+                }
+                overridePendingTransition(0, 0)
+                finish()
+            } else {
+                ToastMessageUtil.showToast(this@NetworkErrorActivity, getString(R.string.network_error_toast))
             }
-            overridePendingTransition(0, 0)
-            finish()
         }
     }
