+++ app/src/main/java/hous/release/android/util/extension/ContextExtension.kt
@@ -1,0 +1,16 @@
+package hous.release.android.util.extension
+
+import android.content.Context
+import android.net.ConnectivityManager
+import android.net.NetworkCapabilities
+
+fun Context.isNetworkConnected(): Boolean {
+    var isConnected = false
+    val cm = this.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
+    val capabilities = cm.getNetworkCapabilities(cm.activeNetwork)
+    if (capabilities != null) {
+        isConnected =
+            capabilities.hasTransport(NetworkCapabilities.TRANSPORT_WIFI) || capabilities.hasTransport(NetworkCapabilities.TRANSPORT_CELLULAR)
+    }
+    return isConnected
+}
